backend/requirements.txt
fastapi
uvicorn
python-jose
pymongo
passlib[bcrypt]
